Connect to a Domo instance

Connect to a Domo instance

# S4 method for DomoDriver
dbConnect(
  drv,
  client_id = domo_client_id(),
  client_secret = domo_client_secret(),
  ...
)

Arguments

drv
an object that inherits from DBIDriver, or an existing DBIConnection object (in order to clone an existing connection).

client_id
OAuth 2.0 client ID as a string.

client_secret
OAuth 2.0 client secret as a string.

...
These dots are for future extensions and must be empty.

Details

By default the client_id and client_secret are read in from environment variables DOMO_CLIENT_ID and DOMO_CLIENT_SECRET, respectively. To obtain the credentials, login to your Domo instance and visit Domo's developer portal. The client must have the "Data" application scope.